only check if promotion is possible once now
